Fix: Improve error messages when resolving tasks
When updating from gulp 3 to 4, this function is the first one to
fail when the file contains the old exec chain:

```
gulp.task('task', ['task2', 'task3'], function() {...});
```

Prior this commit, the error message is too cryptic. This commit fixes
that by specifying in the message, the actual argument it got during
precondition.